refactor: move api error type to separate module
